requirements.txt
# Core dependencies
boto3==1.37.30
dacite==1.9.2
pydub==0.25.1
pyyaml==6.0.2
requests==2.32.3
werkzeug==3.1.3
wget==3.2
pillow==11.1.0
numpy==1.26.4
pydantic==2.7.1
deal==4.24.3
structlog==24.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
prometheus-client==0.19.0

# Development dependencies
tomlkit==0.13.2
ruff==0.3.3
pre-commit==4.2.0
jupyter==1.1.1
notebook==7.3.3
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.0
mypy==1.8.0
black==24.2.0
isort==5.13.2
mutmut==2.4.4
mypyc==1.8.0
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==2.0.0
pip-tools==7.4.0
crosshair-tool==0.0.51

# Security dependencies
bandit==1.7.7
safety==2.3.5
pip-audit==2.7.2

# Performance dependencies
numba==0.59.1
py-spy==0.3.14
memray==1.11.0
pyinstrument==4.6.2

# ML dependencies
scikit-learn==1.4.1.post1
torch==2.2.1
torchvision==0.17.1
scikit-optimize==0.9.0

# Crypto dependencies
pyfhel==3.0.1
phe==1.5.0

# AI Agents dependencies
openai==1.12.0
langchain==0.1.4
langchain-openai==0.0.5

# Hardware integration dependencies (optional)
# qiskit==0.45.1
# cirq==1.2.0
# pennylane==0.32.0
# pyquil==4.0.0
# braket==1.36.1
# qsharp==0.28.302812
# nengo==3.2.0
# brian2==2.5.1
# nengo-dl==3.6.0
# nengo-loihi==1.0.0
# tenseal==0.3.12
